[료칸등록 4단계] 컴포넌트 분리
